Fix SVT-HEVC encode crash when gop_size is 1
SVT-HEVC recommends that "When RateControlMode is VBR, it's the best to
set 'la_depth' to be equal to the intra period value (gop_size)."
'lookahead' concept is to improve the dynamic allocation of p/b frames.
SVT-HEVC does not expect 'la_depth' is set when gop_size is 1 because
it means there'll be I frame only. When 'la_depth' is 1, it causes
SVT-HEVC stack crash.